Se ha añadido para que lance la excepcion de ObjetoNotFound

+++ src/Objeto/Application/Delete/ObjetoDelete.php
@@ -2,6 +2,7 @@
 
 namespace App\Objeto\Application\Delete;
 
+use App\Objeto\Domain\ObjetoNotFound;
 use App\Objeto\Domain\ObjetoRepository;
 
 class ObjetoDelete
@@ -13,6 +14,10 @@ class ObjetoDelete
     public function delete(int $objetoId)
     {
         $objeto = $this->repository->search($objetoId);
+
+        if ($objeto === null) {
+            throw new ObjetoNotFound($objetoId);
+        }
 
         $this->repository->delete($objeto);
     }
